Adding Vulkan
- added vulkan using volk
- cvar system (tbc)

+++ KISS/Code/Common/ProjectDefines.h
@@ -9,7 +9,7 @@
 using byte = uint8_t;
 
 using int32 = int32_t;
-using int32 = int32_t;
+using int64 = int64_t;
 
 using uint32 = uint32_t;
 using uint64 = uint64_t;
